feat: memory func

+++ lib/models/system.dart
@@ -4,7 +4,7 @@ class System {
     required this.memory,
   });
   final List<double> cpuList;
-  final int memory;
+  final double memory;
 
   double get cpuAverage {
     return cpuList.fold<double>(0, (double a, double b) => a + b) /
